refine the exam card file
