--- frontend/styles/card.css
@@ -1,18 +1,0 @@
-.card {
-  background-color: #f0f4f8;
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-  padding: 10px;
-  margin-top: 20px;
-}
-
-.card p {
-  color: #000000;
-}
-
-.card h3 {
-  color: #000000;
-  font-size: larger;
-  font: 1em sans-serif bold;
-}
